# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "EvidenceOut": {
        "properties": {
          "fields": {
            "additionalProperties": true,
            "title": "Fields",
            "type": "object"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "local_id": {
            "title": "Local Id",
            "type": "string"
          },
          "ts": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ts"
          },
          "zeek_uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zeek Uid"
          }
        },
        "required": [
          "local_id",
          "kind",
          "finding_id",
          "zeek_uid",
          "ts",
          "fields"
        ],
        "title": "EvidenceOut",
        "type": "object"
      },
      "EvidencePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EvidenceOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "EvidencePage",
        "type": "object"
      },
      "FeedbackIn": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "enum": [
              "true_positive",
              "false_positive",
              "expected_benign"
            ],
            "title": "Label",
            "type": "string"
          },
          "note": {
            "default": "",
            "maxLength": 2000,
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "label"
        ],
        "title": "FeedbackIn",
        "type": "object"
      },
      "FeedbackOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "label": {
            "enum": [
              "true_positive",
              "false_positive",
              "expected_benign"
            ],
            "title": "Label",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "id",
          "finding_id",
          "label",
          "note",
          "created_at"
        ],
        "title": "FeedbackOut",
        "type": "object"
      },
      "FindingOut": {
        "properties": {
          "benign_causes": {
            "items": {},
            "title": "Benign Causes",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "string"
          },
          "detector_id": {
            "title": "Detector Id",
            "type": "string"
          },
          "detector_version": {
            "title": "Detector Version",
            "type": "string"
          },
          "end_ts": {
            "format": "date-time",
            "title": "End Ts",
            "type": "string"
          },
          "evidence_total": {
            "title": "Evidence Total",
            "type": "integer"
          },
          "feedback": {
            "items": {
              "$ref": "#/components/schemas/FeedbackOut"
            },
            "title": "Feedback",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "local_id": {
            "title": "Local Id",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "primary_entity": {
            "title": "Primary Entity",
            "type": "string"
          },
          "secondary_entities": {
            "items": {},
            "title": "Secondary Entities",
            "type": "array"
          },
          "severity_score": {
            "title": "Severity Score",
            "type": "number"
          },
          "start_ts": {
            "format": "date-time",
            "title": "Start Ts",
            "type": "string"
          },
          "techniques": {
            "items": {
              "$ref": "#/components/schemas/TechniqueOut"
            },
            "title": "Techniques",
            "type": "array"
          },
          "thresholds": {
            "additionalProperties": true,
            "title": "Thresholds",
            "type": "object"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "local_id",
          "type",
          "detector_id",
          "detector_version",
          "primary_entity",
          "secondary_entities",
          "start_ts",
          "end_ts",
          "metrics",
          "thresholds",
          "confidence",
          "severity_score",
          "benign_causes",
          "evidence_total",
          "techniques",
          "feedback"
        ],
        "title": "FindingOut",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentDetailOut": {
        "properties": {
          "cards": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Cards",
            "type": "array"
          },
          "end_ts": {
            "format": "date-time",
            "title": "End Ts",
            "type": "string"
          },
          "finding_count": {
            "title": "Finding Count",
            "type": "integer"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/LinkOut"
            },
            "title": "Links",
            "type": "array"
          },
          "local_id": {
            "title": "Local Id",
            "type": "string"
          },
          "playbooks": {
            "items": {
              "type": "string"
            },
            "title": "Playbooks",
            "type": "array"
          },
          "primary_entity": {
            "title": "Primary Entity",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "severity_label": {
            "title": "Severity Label",
            "type": "string"
          },
          "severity_score": {
            "title": "Severity Score",
            "type": "number"
          },
          "start_ts": {
            "format": "date-time",
            "title": "Start Ts",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "types": {
            "items": {
              "type": "string"
            },
            "title": "Types",
            "type": "array"
          }
        },
        "required": [
          "id",
          "local_id",
          "rank",
          "primary_entity",
          "start_ts",
          "end_ts",
          "severity_score",
          "severity_label",
          "types",
          "finding_count",
          "links",
          "summary",
          "findings",
          "cards",
          "playbooks"
        ],
        "title": "IncidentDetailOut",
        "type": "object"
      },
...
      "LinkOut": {
        "properties": {
          "from_incident": {
            "title": "From Incident",
            "type": "integer"
          },
          "from_local": {
            "title": "From Local",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "to_incident": {
            "title": "To Incident",
            "type": "integer"
          },
          "to_local": {
            "title": "To Local",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "from_incident",
          "to_incident",
          "from_local",
          "to_local",
          "reason"
        ],
        "title": "LinkOut",
        "type": "object"
      },
...
      "NetworkOut": {
        "properties": {
          "allowlist": {
            "additionalProperties": true,
            "title": "Allowlist",
            "type": "object"
          },
          "internal_cidrs": {
            "items": {
              "type": "string"
            },
            "title": "Internal Cidrs",
            "type": "array"
          },
          "known_hosts": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Known Hosts",
            "type": "object"
          }
        },
        "required": [
          "internal_cidrs",
          "known_hosts",
          "allowlist"
        ],
        "title": "NetworkOut",
        "type": "object"
      },
...
      "SliceAccepted": {
        "properties": {
          "slice_id": {
            "title": "Slice Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "slice_id",
          "status"
        ],
        "title": "SliceAccepted",
        "type": "object"
      },
      "SliceOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finding_id": {
            "title": "Finding Id",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "packets": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Packets"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "finding_id",
          "status",
          "size_bytes",
          "packets",
          "error",
          "created_at"
        ],
        "title": "SliceOut",
        "type": "object"
      },
      "TechniqueOut": {
        "properties": {
          "phrase": {
            "title": "Phrase",
            "type": "string"
          },
          "technique_id": {
            "title": "Technique Id",
            "type": "string"
          }
        },
        "required": [
          "technique_id",
          "phrase"
        ],
        "title": "TechniqueOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Tracewright",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/config/network": {
      "get": {
        "operationId": "get_network_api_v1_config_network_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NetworkOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Network"
      }
    },
    "/api/v1/findings/{finding_id}/feedback": {
      "post": {
        "operationId": "add_feedback_api_v1_findings__finding_id__feedback_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Feedback"
      }
    },
    "/api/v1/findings/{finding_id}/slice": {
      "post": {
        "operationId": "create_slice_api_v1_findings__finding_id__slice_post",
        "parameters": [
          {
            "in": "path",
            "name": "finding_id",
            "required": true,
            "schema": {
              "title": "Finding Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SliceAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Slice"
      }
    },
...
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident"
      }
    },
    "/api/v1/incidents/{incident_id}/evidence": {
      "get": {
        "operationId": "get_evidence_api_v1_incidents__incident_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "finding_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Finding Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidencePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence"
      }
    },
...
    "/api/v1/incidents/{incident_id}/report": {
      "get": {
        "operationId": "get_report_api_v1_incidents__incident_id__report_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "md",
              "enum": [
                "md",
                "html"
              ],
              "title": "Format",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report"
      }
    },
...
    "/api/v1/slices/{slice_id}": {
      "get": {
        "operationId": "get_slice_api_v1_slices__slice_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "slice_id",
            "required": true,
            "schema": {
              "title": "Slice Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SliceOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Slice"
      }
    },
    "/api/v1/slices/{slice_id}/download": {
      "get": {
        "operationId": "download_slice_api_v1_slices__slice_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "slice_id",
            "required": true,
            "schema": {
              "title": "Slice Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Slice"
      }
    }
  }
}